+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Categories for each release: Added, Changed, Deprecated, Removed, Fixed, Securit
 
 * New argument '--instance-type-by-executable' arg for `dx run` and `dx-jobutil-job-new`
 * Added new return fields from `dx describe {job/analysis}-xxxx` with `--verbose` argument: 'runSystemRequirements', 'runSystemRequirementsByExecutable', 'mergedSystemRequirementsByExecutable', 'runStageSystemRequirements'
+* `--metrics top` for `dx watch`
 
 ## [352.0] - beta